Fix single precision float bug
This fixes the NSMBU crash. According to both Wiibrew and 6xx_pem.pdf single precision floats are converted to double when they are stored. In the emulator code, they're stored twice though, as if they were paired single instructions. lfs and similar instructions seem to store them into the the register the wrong way too. That will probably have to be fixed as well.