Clean up alarm handling code slighty.
Made the alarm triggering behaviour slightly more clear
by moving it out of a separate function.  We now update
the CPU alarms within the sAlarmLock just in case of races.
Lastly, we no longer update the CPU alarms after an alarm
is cancelled, since in the future our alarm update code can
just store the last CPU alarm that was set and only update
if the new alarm is sooner.  This will save some CPU.