Use a mutex to guarentee thread-safe reservations.
Previously, reservations were not atomically handled, PPC does
not require that the reservation address match the storage
address, this means a compare_and_exchange is not possible.
We have reverted to using a lock for this case.  Note that the
AtomicityLock class is templated to remove all locking overhead
from non-atomic instructions in spite of using a generic
function to handle most load and store operations.