Significantly improved kernel context-switching.
This patch brings two major changes to our handling of context
switching.  The first is that we now switch context's based on
soely the OSContext.  This allows the kernel to be get closer
to no longer depending on coreinit.  I have also updated the
behaviour surrounding the actual host platform fiber switch so
that the semantics are easier to follow.  Lots of comments too!