New simpler shader decompiler.
The previous shader decompiler tried to be very clever in reversing control flow,
that was a bit unreliable as it made a few assumptions which turned out to not
always be true. I believe in the future we could still write a proper decompiler
which correctly reverses control flow by inspecting the stack, but for now let's
keep it very simple and use a basic bitch decompiler which uses an actual shader
stack similar to how the GPU itself works.