Fix bug in attribute buffer divisor selection.
The previous code assumed that VAO's specified their divisor info
on a per-attribute basis when it is actually per-buffer.  We now
track the divisor information for each buffer and set it once at
the end, we also track to ensure that all attributes of a single
buffer use the same divisor information.