Update loader to use a custom memory space for loading.
This is required as some games stupidly expect that the memory
in their MEM2 bucket is zerod out when the application starts
up.  When the loader runs and uses MEM2 for its loader temp
region, this "dirties" that memory, causing those games to fail.
We can't use the system heap as it is not large enough for some
games temporary data (seen up to 0x02900000 bytes).  Note that
we do not keep the loader memory mapped at all times, only when
the loader is actively running.