common: Fix bitfield compilation on GCC.
GCC is a bit funny here, fmt::format must be taking a const T& which
ends up requiring a proper definition of MaxValue in a .cpp somewhere
so it can take address. So to get around that we use static_cast to
force an lvalue to be passed in instead.

Thanks gcc!