Implement GLSL attribute decoding and attrib buffer invalidation.
This commit updates attribute decoding to occur entirely within the
vertex shader in GLSL.  This allows us the flexibility to support
various formats which GL does not, but latte does.  The other major
benefit is that we are able to perform trivial state-tracking of
attribute buffers and avoid swapping and uploading large buffers
when they have not actually changed.