opengl: Fix failure to reuse color buffers as textures.
Color buffer sizes can only be specified to tile granularity, while
texture sizes can be specified to texel granularity.  If a game renders
to an intermediate framebuffer with a non-tile-aligned effective size,
then sets that color buffer as a texture, the texture would be detected
as a separate surface due to the size difference, preventing the color
buffer's contents from being rendered.  (For example, Xenoblade uses
intermediate framebuffers of effective size 768x620 and 384x310, but
the framebuffer sizes are set to 768x624 and 384x320 respectively.)

This change allows surfaces to be resized as long as the guest surface
pitch remains unchanged.  To deal with differing sizes, texture
sampling calls multiply the texture coordinate with a scale factor
computed from the texture size specified by the program and the actual
OpenGL surface size.