Implement better multi-use texture semantics.
We not maintain a list of textures for each PPC texture that
exists in memory.  This allows us to use properly sized texture
objects on the host-GPU while also supporting various ways that
games could (miss-)use texture data...  Note that this commit
mostly reverts commit (9674136) and also cleans up the handling
of different data formats to a common path.