Implement a circular buffer to handle command buffer pool.
This more closely mimics the behaviour present in GX2.  Note that
we currently require a lock to handle the pool itself, though in
the future, we may be able to implement a lock-free queue which
is used by the GPU to store `completed` buffers in, which can
later be picked up by GX2 when it tries to do an allocation. This
commit also cleans up a few things that were wierd about our
display list implementation.