opengl: Add workaround for NVIDIA GLSL compiler bug.
The NVIDIA GLSL compiler (as of driver version 367.35) incorrectly
reports a "stride mismatch" error on a shader declaring multiple
transform feedback buffers with different strides if the xfb_buffer
layout qualifier is specified in the definition of the buffer itself.
The compiler works correctly if we first change the default buffer
index and then define the buffer without the xfb_buffer qualifier.

The bug has been reported to NVIDIA as bug #160816-000177.