Include isDepth in host surface selection.
I'm not actually sure if this is neccessary, and we should
remove the isDepthBuffer stuff from everwhere it is currently
if we are able to.  I've heard that you can read
GL_DEPTH_COMPONENT32F as a normal texture without needing to
do a conversion from various sources, but I have not yet
confirmed if this is the case.  For now lets work
(potentially) suboptimally by doing the copy so we at least
know it will behave properly.