opengl: Invert handling of ZCLIP_NEAR_DISABLE in PA_CL_CLIP_CNTL.
In OpenGL, depth "clamping" is the opposite of clipping: when depth
clamping is enabled, the near and far clip planes are _dis_abled,
allowing all depth values to pass through to the final depth buffer
store; at that point, the depth value is clamped to the range [0,1].
This behavior corresponds to disabling Z-clipping with the
ZCLIP_NEAR_DISABLE bit.

Also add an assertion that the near and far clipping flags have the
same state, since OpenGL does not allow clamping on one side only.