Revert: opengl: Handle VGT_MULTI_PRIM_IB_RESET_EN.
This breaks quite a bit of things, I think there is more to
this than that.  On GX2, primitive restart appears to always
be enabled, with the specific reset value being controlled
rather than the feature being turned on/off entirely.  Note
that the register you are checking is actually never touched
by gx2.rpl.