opengl: Disable early fragment tests for shaders with discard enabled.
Early fragment tests are mostly an optimization, but they do change
behavior with respect to pixels which are discarded by the fragment
shader.  The DB_SHADER_CONTROL EARLY_Z flag simply means "use early
fragment tests when possible", so we shouldn't enable them when they
would change shader behavior.

It seems most likely that the hardware uses the KILL_ENABLE flag to
decide whether to suppress EARLY_Z for this reason, so do the same
here, but warn if a shader uses KILL instructions and fails to set
KILL_ENABLE.

Also lower the level of the "early-z disabled" log message from
warning to debug, since it seems to be common to leave the EARLY_Z
flag enabled regardless of whether it has an effect.