gpu: Remove flags from EVENT_WRITE high address word.
At minimum, the word does not have a DATA_SEL field, since GX2 query
functions pass zero (DISCARD) in those bits and expect the result to
not be discarded.  Removed INT_SEL from that word as well since there
is no evidence for its existence (and it's not handled anyway).