opengl: Only upload uniform registers which have been recently written.
"Recently" is defined by an update generation counter; the counter is
stored with each uniform(*) when one of its component registers is
written, and is incremented and saved with a shader after uploading
changed uniforms, where "changed" is detected as a uniform whose last
write generation is not less than the shader's last upload generation.

(*) To be precise, uniform change state is tracked in groups of 16
uniform vectors (64 registers), as a balance between the cost of
uploading too much unchanged data to the host GPU and the cost of
checking the changed status of each uniform individually.

In addition to reducing the amount of data uploaded to the host GPU,
this typically reduces the number of GL calls per frame by 5-10% (due
to cases where no uniforms have changed at all), and as a bonus it
significantly improves the replay speed of GL trace files.