Correct issue with interrupt handling across thread switches.
True coreinit.rpl behaviour is that interrupt are disabled whenever
there is a scheduler lock active.  We have elided this due to our
C++ code being uninteruptible, but this causes issues when games
already have interrupts disabled and then do a rescheduling
operation.  We have to force-enable them for the switch instead.