Implement synchronization of the guest and host GPUs.
This commit will allow us to rewrite some of the streamout
and other logic such that it no longer blocks the host GPU
directly, but only blocks PPC waiting for stuff.  Note that
I originally had it checking the sync objects after every
pm4 command, but this had a lot of overhead, and I don't
believe it will make any real difference.