gx2: Only warn when pointer passed to end display list and begin display list differ.
FAST Racing Neo seems to do this, the pointer passed to GX2EndDisplayList
isn't really used for anything in gx2.rpl (other than something related to
pm4 capture, which has no affect on how the game renders).

I think this is likely to be an application bug rather than one of ours.