coreinit: Allocate memory heap pointers passed to __preinit_user from heap.
Apparently games can store these pointers and use them later in the game,
Yoshi's Wooly World exhibits this behaviour.

This means we cannot use a stack object like previously, because the stack
object was being freed and is no longer valid. We could move the stack
object up one scope and it would persist as long as the game's main thread
did, but it feels safer to just have it on the system heap.