opengl: Use a sorted resource map to optimize resource updates.
Since resources are not currently expunged when the memory is reused for
another purpose, the resource maps grow over time, causing the linear
scans in notify{Cpu,Gpu}Flush() and surfaceSync() to consume more and
more CPU as time goes on.  This change introduces a ResourceMemoryMap
class to keep a sorted list of resource memory endpoints, so the scans
can touch exactly those resources that fit the requested memory range
without having to iterate over every single allocated resource.