espresso: Ignore register fields which are ignored by hardware.
A real Espresso ignores some register fields which are unused by certain
instructions, such as frC in fadd.  Such invalid opcodes are presumably
not generated by any real-life compilers, but ignore such fields anyway
for completeness.  This may also provide a marginal performance boost to
the interpreter.

This is probably not a complete set of ignored fields, but it's enough
to pass all relevant tests in the current test suite at
<http://achurch.org/cpu-tests/ppc750cl.s>.