loader: Do not use virtual loader heap during loading.
Our virtual loader heap is stealing from IOS memory heap locations, so I
need to remove the loader heap.

On a console they have a small loader heap and the loader only reads
chunks at a time to prevent filling the heap. Our loader is loading
everything at the same time which is why we expanded the loader heap to
be so large.